services/telegram_bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram Bot Interface"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "structlog>=25.1.0",
    "redis[hiredis]>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
